Remove service-desk related modules and move the content into the `servicedesk` package inside `jira-slack-server-integration-plugin`
