+++ jira-slack-server-integration/jira-slack-server-integration-plugin/pom.xml
@@ -22,11 +22,6 @@
             <artifactId>slack-server-integration-common</artifactId>
         </dependency>
         <dependency>
-            <groupId>com.atlassian.plugins</groupId>
-            <artifactId>jira-service-desk-compat-main</artifactId>
-            <version>${slack.common.version}</version>
-        </dependency>
-        <dependency>
             <groupId>com.github.seratch</groupId>
             <artifactId>jslack</artifactId>
         </dependency>
@@ -80,6 +75,11 @@
         <dependency>
             <groupId>com.atlassian.jira</groupId>
             <artifactId>jira-admin-project-config-plugin</artifactId>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.atlassian.servicedesk</groupId>
+            <artifactId>jira-servicedesk-api</artifactId>
             <scope>provided</scope>
         </dependency>
         <dependency>
