+++ jira-slack-server-integration/jira-slack-server-integration-plugin/src/test/java/com/atlassian/jira/plugins/slack/servicedesk/ServiceDeskCompatibilityDispatcherTest.java
@@ -1,17 +1,17 @@
-package com.atlassian.jira.plugins.slack.compat;
+package com.atlassian.jira.plugins.slack.servicedesk;
 
 import org.junit.Before;
 import org.junit.Test;
 
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.sameInstance;
-import static org.junit.Assert.assertThat;
 
 public class ServiceDeskCompatibilityDispatcherTest {
     ServiceDeskCompatibilityDispatcher target;
 
     @Before
-    public void setUp() throws Exception {
+    public void setUp() {
         target = new ServiceDeskCompatibilityDispatcher();
     }
 
@@ -22,6 +22,6 @@ public class ServiceDeskCompatibilityDispatcherTest {
 
     @Test
     public void getHelperReturnsNonEmptyHelper() {
-        assertThat(target.getHelper().get(), sameInstance(ServiceDesk4CompatibilityHelper.getInstance()));
+        assertThat(target.getHelper().get(), sameInstance(DefaultServiceDeskCompatibilityHelper.getInstance()));
     }
 }
