+++ jira-slack-server-integration/pom.xml
@@ -14,7 +14,6 @@
 
     <modules>
         <module>jira-slack-server-integration-plugin</module>
-        <module>jira-service-desk-compat</module>
     </modules>
 
     <properties>
@@ -108,6 +107,11 @@
                         <groupId>com.atlassian.plugin</groupId>
                     </exclusion>
                 </exclusions>
+            </dependency>
+            <dependency>
+                <groupId>com.atlassian.servicedesk</groupId>
+                <artifactId>jira-servicedesk-api</artifactId>
+                <version>${jira.version}</version>
             </dependency>
             <dependency>
                 <groupId>com.atlassian.jira</groupId>
